refactor(chat-store): hoist extractToolErrorMessage to shared mcp-utils
Both stores called identical 10-line copies of the helper. Future
divergence (e.g. a richer MCP error shape) would silently inconsistify
the two. New module lives next to its callers; only this helper for now.
